finish animation, reorganize
