+++ .gitignore
@@ -12,7 +12,7 @@
 /build
 
 # keys
-/src/keys.js
+/src/utils/keys.js
 
 # misc
 .DS_Store
