+++ src/components/ChapterStack.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Stack, Box } from "@mui/material";
+import { Box } from "@mui/material";
 import NextPrevButtons from "./NextPrevButtons";
 
 const ChapterStack = ({
